separate logic for proc and auto extraction
